GrAna: Small fixes.
Changed an analysis ID to be able to reference it from GrAna text. Also
fixed a casting problem in AspectRatioAnalysis.

+++ plugins/de.cau.cs.kieler.grana/src/de/cau/cs/kieler/grana/analyses/AspectRatioAnalysis.java
@@ -49,8 +49,8 @@ public class AspectRatioAnalysis implements IAnalysis {
         
         // Compute the aspect ratio
         Object[] areaResults = (Object[]) o;
-        double width = (Float) areaResults[AreaAnalysis.INDEX_WIDTH];
-        double height = (Float) areaResults[AreaAnalysis.INDEX_HEIGHT];
+        double width = (Double) areaResults[AreaAnalysis.INDEX_WIDTH];
+        double height = (Double) areaResults[AreaAnalysis.INDEX_HEIGHT];
         
         double aspect = width / height;
